Move some functionality into modules. Closes #36
This is just a small granularization. Even more methods may be exported into modules.
All the functions that require instance object remain in main file.
Mouse events handlers may be moved into a separate module but in this case either
methods should be called with plugin instance context, or plugin instance should be
passed as an argument.