Add namespaces to events. Move API creation into a module
Public API and each module API methods are namespaced.
This means if you call a plugin API method that results into a panzoom event
then this event will contain that plugin name as event namespace.

Such a behaviour is useful for plugins that want to cancel certain events and
trigger those events by themselves without getting into a recursion. One such
example is a plugin that will animate pan/zoom.

API creation was moved into a module. Reference to inner methods is intentionally
hidden so that it is not accessible from exterior.

Related to #98